polypoly_func/requirements.txt
azure-functions
azure-identity
azure-storage-file-datalake>=12.14.0
databricks-sdk>=0.20.0
requests>=2.31.0
python-dotenv>=1.0.0
